+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.1.6
 * Added running tests in specific order.
+* Changed interface of ErrorBoard. Refactored handling of command line arguments.
 * Changed displaying error details in ErrorBoard. Now it requires `--error-details` option to be available.
 * Changed displaying command line argument errors without trace for test.
 * Changed paths to files and folder to have same format. Windows format with `single backslash` as separator.